% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qidData.R
\docType{data}
\name{QID_exampleData}
\alias{QID_exampleData}
\title{Data for reproducing results reported in "Quantifying Interpersonal Dynamics".}
\format{An object of class \code{data.frame} with 73416 rows and 7 columns.}
\usage{
data(QID_ExampleData)
}
\description{
A dataset containing a minimal set of variables for demonstrating rties analyses.
}
\keyword{datasets}
